fix: PP not found error due to @Cachable annotation in ProtocolParamService.getProtocolParam method and refactoring
