+++ stores/epoch/src/main/java/com/bloxbean/cardano/yaci/store/epoch/service/ProtocolParamService.java
@@ -5,7 +5,6 @@ import com.bloxbean.cardano.yaci.store.epoch.domain.EpochParam;
 import com.bloxbean.cardano.yaci.store.epoch.storage.EpochParamStorage;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -16,7 +15,6 @@ import java.util.Optional;
 public class ProtocolParamService {
     private final EpochParamStorage epochParamStorage;
 
-    @Cacheable("protocolParams")
     public Optional<ProtocolParams> getProtocolParam(int epoch) {
         return epochParamStorage.getProtocolParams(epoch)
                 .map(EpochParam::getParams);
